Simplify endpoint organization
